+++ 3542/2/commands/checkout.rb
@@ -1,4 +1,4 @@
-require './validation/validation'
+require './validations/validation'
 require './confirmation/confirmation'
 
 module Checkout
