+++ 3542/2/validations/checkin_helper.rb
@@ -1,0 +1,11 @@
+module Validation
+  module CheckinHelper
+    def checkin?
+      chat_session[session_key]['checkin']
+    end
+
+    def not_checkin?
+      !checkin?
+    end
+  end
+end
