+++ 3542/2/validations/registration.rb
@@ -1,0 +1,7 @@
+module Validation
+  module Registration
+    def registered?
+      chat_session[session_key]
+    end
+  end
+end
