+++ 3542/2/validations/selfie.rb
@@ -1,0 +1,7 @@
+module Validation
+  module Selfie
+    def selfie?
+      payload['photo']
+    end
+  end
+end
